+++ changelog.md
@@ -4,6 +4,7 @@
 ### features
 - Add Smart Auto Hide for Seelenweg.
 - Add visible Separators Option
+- Enable animations for items into LEFT, TOP, RIGHT positions
 
 ### enhancements
 - Now the copy handles option will return hexadecimal handles instead decimal (good for faster debug in tools like spy++).